Add /usr/sbin/dmidecode as a dummy command
A component of aTrust, /usr/share/sangfor/aTrust/resources/bin/dmidecode_processor_tool,
seems to search for and try to run dmidecode in /usr/{bin,sbin}.
Exactly, it runs `sh -c 'path-to-dmidecode -t processor'` periodically
(or it is called periodically). If there is no dmidecode command in
/usr/bin or /usr/sbin, it runs `sh -c '-t processor'` instead, which
throws the `sh: 1: -t: not found` spam in console.

This commit creates a symbolic link to /bin/false in /usr/sbin/dmidecode
to suppress the spam.